Extract Firecracker fork snapshot prep
